requirements.txt
streamlit>=1.28
streamlit-extras
langchain>=0.0.217
langchain_openai
openai>=1.2
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3
streamlit-feedback
python-dotenv
